requirements.txt
fastapi
jinja2
python-multipart
uvicorn
pyjwt
pyodbc
docx-mailmerge
pandas
openpyxl
python-dotenv
pyotp